Core and OS baseline insertion benchmarks
