+++ test/CMakeLists.txt
@@ -10,6 +10,7 @@ add_custom_target(benchmarks)
 
 add_subdirectory(benchmark-common-tasks)
 add_subdirectory(benchmark-crud)
+add_subdirectory(benchmark-larger)
 # FIXME: Add other benchmarks
 
 set(CORE_TEST_SOURCES
