+++ test/benchmark-larger/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(realm-benchmark-larger main.cpp)
+add_dependencies(benchmarks realm-benchmark-larger)
+target_link_libraries(realm-benchmark-larger TestUtil)
