+++ test/object-store/CMakeLists.txt
@@ -134,3 +134,4 @@ endif()
 
 add_subdirectory(notifications-fuzzer)
 add_subdirectory(benchmarks)
+add_subdirectory(benchmarks-larger)
